requirements.txt
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.7.0
